Add ListenableCache backed by Caffeine and WeakRefs
This allows to have size-limited caches in cases where SoftRefs are too
optimistic.

The surface to Caffeine cache is very small, only getIfPresent() and put()
are used. This should make it very easy to offer Guava etc as alternative
options. I decided for Caffeine now, because it is favored by benchmarks I
could find, and also because it is a smaller dependency.