Add CellLoader for filling cells of type T with data
This is an alternative to CacheLoader<Long,Cell<A>> and unlike that,
CellLoader is not typed on the access, allowing to use the same CellLoader
for all volatile/dirty access versions.

LoadedCellCacheLoader.get(...) creates a CacheLoader<Long,Cell<A>> of the
appropriate access type, that is backed by a CellLoader.