Trigger CacheRemover.invalidate() etc from Cache.invalidate() and clarify javadoc.
This only makes sense for caches that have a global CacheRemover (instead of
per entry). Otherwise, behaviour wrt the CacheRemover would depend on
whether the entry is in the cache, which is usually not what is desired.
For example, invalidateAll() is expected to remove clear the backing disk
cache, in particular also for entries that are no longer in the memory
cache. With per-entry removers, there is no way to achieve that.