Add IoSync.shutdown method
Also add IoSync as member to DiskCachedCellImg so it can be shutdown as well.

Fixes #12

Fun fact: I was unable to create a `Writer[]` array (Generic array creation error) and was very confused by it. Now I understand that it is because it's an inner class of `IoSync`, which is generic and thus, Writer is generic, as well.